I decided do commit this without having it tested with fallout1 data files.
I am 95% sure that it should work without a glitch,
but there is the posibility ... you know.

On the light side, this should speed up
startup significantly, as the long
time spend in the constructor of
fallout::DAT2 is now spread over the first
8 seconds or so...

You will _not_ notice any improvement,
in case you're trying to load a file that does not
exist right at startup. This is the case if the
BackgroundMusic file, set in fife.config
is not found.