* guichan 0.6 support
You have to define the version by using "scons guichan=X" where X is
the sub-version number of the guichan version you want to use; this
_has_ to be correct, no testing is done so far.

Possible values are: 4, 5, 6 for the respective guichan releases; the
default is still 4.

0.6 seems to work fine (so far).

I suggest to skip 0.5; if you want to use it you have to modify 
src/basiccontainer.cpp and insert "return;" after line 303.