Working on object ownership.
This will disallow adding objects to different
inventories or adding an object to the inventory
of an object in its inventory.

Although there are use cases for this,
we will disallow cycles, cause these
would create memory leaks with the current design.

Also (but this is not yet implemented) 
an objects location will be coupled with the
layer it is located on. Setting the location
of an object in an inventory does not make
a lot of sense.

An unit test was written for the objects
ownership semantics.