This is a large commit, (mostly) putting the map model into shape. In no special order:
Objects ownership is now tracked either by another
object (inventory) or an layer. In this iteration
I had to leave object.elevation/layer as numbers,
which will break in case you insert a layer in a
map (so that ids are garbled).

There's a test_case for the objects ownership,
which runs successfully.

The 'map/structure.*' files have been removed and
the fallout loader just uses an own createElevation
function, instead of littering the creation of
elevations.

Renamed the cameras zoomTo to moveTo and moveTo is
now called jumpTo.

Some more map functions are exported, but this is
still untested.

The test_events profiling test case has been disabled,
as it really takes an annoying time to run.

The smartlist template class has just been written
to a simple object list class.

A lot of performance tweaking might be in order, once
everything works correctly, as especially Layer
and ObjectList execute unnecessary slow functions.