This unifies the map model classes.
Map and Elevation now have insert/add/remove
functions, these are exported to lua and
tested with a new test case.

Duplicate entries are not allowed,
that is an elevation can only be
on one map, a layer only on one elevation.

Other smaller changes were made while i was at it:
* Layer::create throws on invalid sizes
* New exception class 'Duplicate'
* Deletion of attributes