Incorporated Tim Goyas fine Fake-alpha-channel remover for SDL images.
Performance upgrade is questionable in a way, as the displayed framerate
did not change significantly in my case (11+/-2 fps).
I have checked the actually optimized code path
in Control::update - and there i had numbers changing.

30..40 msec per rendering of the map
versus 15..25 with the optimizer used. Nice!

Since the optimizer might fail to recognize a correctly
used alpha channel I have added a boolean config option
SDLRemoveFakeAlpha=yes.

Also note that this change does nothing in the OpenGL
case (where it would be useless) and it is also only
enabled for the ImageProvider for now.

Thanks again to Tim Goya.