* This fixes the case were the engine would clear the screen twice
The editor and demos now suffer from a problem where if a map is not loaded the screen doesnt get cleared and you see traces of the UI from previous frames.  For now the editor, RPG demo and Shooter force the engine to clear the screen every frame.  This is not ideal as once a map is loaded they will be clearing the screen twice per frame.