* Fixed the show/hide of the toolbar and toolbox (which suffered from the same problem) in the editor.
Some problems still exist with the toolbars.  When undocking a toolbar it cuts of the bottom half of the icons/buttons which gets fixed if you hide and show the toolbar.  

Another issue is the orientation of the toolbar gets set and saved to either horizontal or vertical based on which dockarea to attach it to.  This may not be the desired outcome. 

fixes[t:635] and fixes[t:640]