Use integer divisions in some pychan calculations
All size and position attributes are integers, so this is one of the
smallest steps towards python3 support of any kind. It also is one of
the lowest-hanging.

Fixes some coding style issues related to whitespace on the way, and
clarifies some operations as well by adding parens (keeping semantics).