Reverted some changes.  Version info comes from version.h again.
I changed it back because it was working.  The only variable that
gets injected is the git hash.  The hash is entirely optional
and automatically gets determined.  You can disable it with a scons
option (--without-githash).  This option will be useful to you if
you are working on fife and compiling it multiple times between
multiple commits because each time the hash changes you need to
compile everything!

refs #769