Removed the lighting model setting from the default settings window
This prevents clients from having to specify a lighting model in their
settings gui even though the client doesn't make use of lighting.  This is
hardly an ideal solution.  Client's are still required to specify several other
widgets.  We should probably not provide a default settings window and
force clients to specify one.  This of course would require us to specify
a setting window for each demo that uses one but it's not that big of a
deal.

fixes #781