Don't allow "in-source-tree" build directory for older cmake
In older versions (<3.0) of CMake, the UseSWIG.cmake module does some
matching against the SOURCE_DIR and then against the BINARY_DIR. If the
BIN_DIR has the SRC_DIR as a parent, the paths of certain files in the
generated Makefile might end up wrong and the files cannot be compiled.

With this patch, cmake does not allow the BINARY_DIR to have the
SOURCE_DIR as a parent for CMake versions < 3. Instead, it shows a hint
and exits.